Tweak contract on make-heap's predicate argument
The new contract can be implemented much more efficiently than the
contract it replaces in the case that we have an n-ary procedure that
is known to return a single result, e.g., <. In that case, the
new contract is just a first-order check, but the old contract:

  (-> any/c any/c any/c)

has to create a chaperone. This makes a difference in some uses of
the data/heap library, since it ends up making a lot of calls to that
function. In particular, the code below gets about a 4x speedup

  #lang racket/base
  (require data/heap)

  (define (heap-sort ns)
    (define h (make-heap <))
    (for ([x (in-list ns)])
      (heap-add! h x))
    (let loop ()
      (cond
        [(= 0 (heap-count h)) '()]
        [else
         (define n (heap-min h))
         (heap-remove-min! h)
         (cons n (loop))])))

  (collect-garbage)
  (define l '(2 4 6 8 10 12 14 16 18 20 22 24 26 28 30
                29 27 25 23 21 19 17 15 13 11 9 7 5 3 1))
  (time
   (for ([x (in-range 10000)])
     (heap-sort l)))